+++ RAnalogs/RAnEn/R/boot.fun.ver.R
@@ -18,7 +18,7 @@ boot.fun.ver <- function( temp, R, fun = function(d,i){ mean( d[i,1] ) }) {
   temp <- data.frame( val = na.omit( temp ) )
   boot <- rep(NA,3)
 
-  stopifnot(require(boot))
+  stopifnot(requireNamespace(boot, quietly = T))
   
   if( nrow(temp) > 1 ) {
     err.boot   <- boot::boot(temp, fun, R=R)