+++ RAnalogs/RAnEn/R/formatObservations.R
@@ -103,7 +103,7 @@ formatObservations <- function(
   circular.pars = NA, col.station.name = NA,
   show.progress = F) {
   
-  stopifnot(require(dplyr))
+  stopifnot(requireNamespace(dplyr, quietly = T))
   
   # Sanity check
   stopifnot(is.data.frame(df))