+++ RAnalogs/RAnEn/R/writeConfiguration.R
@@ -33,7 +33,7 @@
 writeConfiguration <- function(config, obs.par.names, fcsts.par.names,
                         xs, ys, file.prefix = '', folder = './',
                         silent = F) {
-  stopifnot(require(ncdf4))
+  stopifnot(requireNamespace(ncdf4, quietly = T))
   
   # Sanity checks
   stopifnot(length(obs.par.names) == dim(config$search_observations)[1])