+++ NEWS.md
@@ -3,6 +3,7 @@
 - Negative values in indices were removed by default. Added a parameter `debug` for backward compatibility.
 - Added function `RAnEn::formatObservations`
 - Updated the docker image
+- Change `require` to `requireNamespace` to avoid function name conflicts
 
 # PAnEn 3.6.5